changed every file, debugged many things

+++ vec.h
@@ -12,7 +12,7 @@ void init_vec(vec *v);  //initialises the vector
 int resize_vec(vec *v, int c);  
 void pb_vec(vec *v, double *val);   //add a pair of integers to the vector
 void set_vec(vec *v, double *val);  //updates value of a present pair
-double get_vec(vec *v, int in);   //returns the second element of the pair if we pass the first element
+double get_vec(vec *v, double in);   //returns the second element of the pair if we pass the first element
 void del_vec(vec *v, double ToDelete);    //deletes vector element based on first element 
 int fr_vec(vec *v);
